ex: move sem tree render to own file
